Remove the notion of 'network' entirely
Money-Tree, in essence is a key-derivation tool.
The keys it derives are not aware of the 'coin',
or 'version'. So there is now no notion of network
held as state on anything.

Money-Tree derives keys, and in the various forms of serialization
you can specify which type of coin (version bytes) you'd like to use.

Although it may seem tedious, it allows trees to be
coin-agnostic until serialization (and theoretically
the usage of the same key-pairs for different coin
addresses.)